Remove legacy CSS mixin tests
